Project setting <Ver1.1>
add Project directory or Flask setting
